Other SEO fixes
Add lang attribute
Remove ratings for now. It must have values baked in during the build.
So during the build I need to connect to the DB and get the results.
That will be a good improvement overall because cold start of the server
takes about a minute and user sessions last for 15 seconds.